requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.1
openai>=1.0.0
gunicorn>=20.1.0
google-generativeai>=0.8.0,<1.0.0
SQLAlchemy>=1.4.0
rich>=13.0.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
redis>=4.5.0
alembic>=1.12.0
nest_asyncio>=1.5.6